backend/requirements.txt
faiss-cpu
Flask
Flask-Cors
google-generativeai
numpy
pandas
python-dotenv
sentence-transformers